+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Updated minimum supported PHP version to 7.4
 ### Removed:
   - Usage of `Api:api` with not fully qualified class name
+  - Usage of `Api::HTTP_RESPONSE_*` constants
 
 ## 2.2.0 / Unreleased
 ### Added: