+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Usage of `Api::HTTP_RESPONSE_*` constants
   - Usage of response format other than `json`
   - Passing a string as `$params` to the methods `create` and `update` of the `PullRequests`, `BranchRestrictions`, `Pipelines`, and `Repository` API
+  - `ClientInterface`, use `Client` instead
 
 ## 2.1.0 / 2021-07-23