+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   - Usage of response format other than `json`
   - Passing a string as `$params` to the methods `create` and `update` of the `PullRequests`, `BranchRestrictions`, `Pipelines`, and `Repository` API
   - `ClientInterface`, use `Client` instead
+  - Passing a string as `$params` to the `request` method of `Client`
 
 ## 2.1.0 / 2021-07-23